Creación de modelo, repo, servicio y controlador para Producto

+++ proyectoFormativo2-master/src/main/java/com/example/proyectoFormativo/Repository/IHistoriaMedicaRepository.java
@@ -1,5 +1,6 @@
 package com.example.proyectoFormativo.Repository;
 
+
 import com.example.proyectoFormativo.Model.HistoriaMedica;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
@@ -7,12 +8,9 @@ import org.springframework.stereotype.Repository;
 import java.util.Optional;
 
 @Repository
-public interface IHistoriaMedicaRepository extends JpaRepository<HistoriaMedica, Integer> {
+public interface IHistoriaMedicaRepository extends JpaRepository<HistoriaMedica, Long> {
 
-    /**
-     * Busca un historial médico usando el ID de la mascota.
-     * Esto es crucial para la vista del veterinario.
-     */
-    Optional<HistoriaMedica> findByMascota_Id(Integer mascotaId);
+
+    Optional<HistoriaMedica> findByMascota_Id(Long mascotaId);
 
 }
